# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: LazyCloud API
  version: "1"
servers:
  - url: /
security:
  - bearer: []
  - session: []
paths:
...
  /v1/workspaces/{workspace}/apps/{app}/endpoints/{endpoint}:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/AppPath"
      - $ref: "#/components/parameters/EndpointPath"
    get:
      operationId: getEndpoint
      summary: An endpoint, its active release and the URLs it answers on
      parameters:
        - $ref: "#/components/parameters/VersionQuery"
      responses:
        "200":
          description: The endpoint
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HttpWorkload" }
        default: { $ref: "#/components/responses/Error" }
  /v1/workspaces/{workspace}/apps/{app}/asgi/{endpoint}:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/AppPath"
      - $ref: "#/components/parameters/EndpointPath"
    get:
      operationId: getAsgi
      summary: An ASGI or realtime app, its active release and the URLs it answers on
      parameters:
        - $ref: "#/components/parameters/VersionQuery"
      responses:
        "200":
          description: The ASGI app
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HttpWorkload" }
        default: { $ref: "#/components/responses/Error" }
  /v1/workspaces/{workspace}/apps/{app}/functions/{function}/invoke:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/AppPath"
      - $ref: "#/components/parameters/FunctionPath"
    post:
      operationId: invokeFunction
      summary: Run the function's active release with JSON arguments
      description: >-
        The body's `args` and `kwargs`, or else the whole body, become the
        arguments, and query parameters are added as keyword arguments.
      parameters:
        - $ref: "#/components/parameters/InvokeWait"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvocationBody" }
      responses:
        "200":
          description: The task, with its result when it succeeded within the wait
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invocation" }
        default: { $ref: "#/components/responses/Error" }
  /v1/workspaces/{workspace}/apps/{app}/functions/{function}/versions/{version}/invoke:
    parameters:
      - $ref: "#/components/parameters/WorkspacePath"
      - $ref: "#/components/parameters/AppPath"
      - $ref: "#/components/parameters/FunctionPath"
      - name: version
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    post:
      operationId: invokeFunctionVersion
      summary: Run one version of the function with JSON arguments
      parameters:
        - $ref: "#/components/parameters/InvokeWait"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InvocationBody" }
      responses:
        "200":
          description: The task, with its result when it succeeded within the wait
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invocation" }
        default: { $ref: "#/components/responses/Error" }
  /v1/domains:
    get:
      operationId: listDomains
      summary: Domains the caller registered, by hostname
      parameters:
        - name: after
          in: query
          description: Return domains whose hostname sorts after this.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: Registered domains
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DomainList" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: registerDomain
      summary: Register a hostname the caller's workspaces may serve from
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DomainRequest" }
      responses:
        "201":
          description: The registration and the DNS records it waits for
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
        default: { $ref: "#/components/responses/Error" }
  /v1/domains/{hostname}:
    parameters:
      - name: hostname
        in: path
        required: true
        schema: { type: string, maxLength: 253 }
    get:
      operationId: getDomain
      summary: One registration, re-read from the provider while it is unsettled
      responses:
        "200":
          description: The registration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: removeDomain
      summary: Retire a registration no deployment serves from
      responses:
        "204":
          description: Removed
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
    session:
      description: >-
        The dashboard's browser session, set by GitHub sign-in. Requests
        that change state with it must send an Origin header naming the
        dashboard.
      type: apiKey
      in: cookie
      name: __Host-lazycloud_session
  parameters:
    WorkspacePath:
      name: workspace
      in: path
      required: true
      schema: { $ref: "#/components/schemas/Name" }
    AppPath:
      name: app
      in: path
      required: true
      schema: { $ref: "#/components/schemas/AppName" }
    FunctionPath:
      name: function
      in: path
      required: true
      schema: { $ref: "#/components/schemas/WorkloadName" }
...
    EndpointPath:
      name: endpoint
      in: path
      required: true
      schema: { $ref: "#/components/schemas/WorkloadName" }
...
    VersionQuery:
      name: version
      in: query
      description: Describe this version instead of the active one.
      schema: { type: integer, minimum: 1 }
    InvokeWait:
      name: wait_seconds
      in: query
      description: Hold the request until the task finishes or this many seconds pass.
      schema: { type: integer, minimum: 0, maximum: 60, default: 60 }
  responses:
    Error:
      description: A typed error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { $ref: "#/components/schemas/ErrorCode" }
        message: { type: string }
    ErrorCode:
      type: string
      enum:
        - invalid_request
        - unauthenticated
        - forbidden
        - not_found
        - conflict
        - payload_too_large
        - unsupported
        - too_many_pending_tasks
        - task_not_finished
        - unavailable
        - internal
    Name:
      type: string
      pattern: "^[a-z][a-z0-9-]{0,62}$"
    AppName:
      type: string
      pattern: "^[a-z][a-z0-9_]{0,62}$"
    WorkloadName:
      type: string
      pattern: "^[A-Za-z_][A-Za-z0-9_-]{0,62}$"
    Sha256:
      description: Lowercase hex SHA-256 digest
      type: string
      pattern: "^[0-9a-f]{64}$"
...
    FunctionSpec:
      type: object
      required: [name, handler, source, image, resources]
      properties:
        name: { $ref: "#/components/schemas/WorkloadName" }
        handler:
          description: "module:qualname within the source archive"
          type: string
          pattern: "^[A-Za-z_][A-Za-z0-9_.]*:[A-Za-z_][A-Za-z0-9_.]*$"
        source:
          $ref: "#/components/schemas/SourceRef"
        image:
          $ref: "#/components/schemas/ImageSpec"
        resources:
          $ref: "#/components/schemas/Resources"
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 86400
          default: 3600
        retry_policy:
          $ref: "#/components/schemas/RetryPolicy"
        concurrency:
          description: Tasks one container runs at once, one process per slot.
          type: integer
          minimum: 1
          maximum: 256
          default: 1
        # Defaults: 10, 0 for a scheduled function, and -1 whenever
        # autoscaler.min_containers is above zero. -1 leaves idle containers
        # above the minimum to the planner, which stops them at once.
        keep_warm_seconds:
          description: Idle seconds before a container above the minimum stops.
          type: integer
          minimum: -1
          maximum: 86400
        autoscaler:
          $ref: "#/components/schemas/Autoscaler"
        max_pending_tasks:
          type: integer
          minimum: 1
          maximum: 1000000
          default: 100
        environment:
          type: object
          additionalProperties: { type: string }
        # Five cron fields, an alias such as @hourly, or `every N` minutes
        # (m, at most 59), hours (h, at most 23) or days (d, at most 31).
        # Deploy stores the normalized expression.
        cron:
          description: Run the function on this UTC schedule.
          type: string
          minLength: 1
          maxLength: 160
        secrets:
          description: Secrets the container receives as environment variables of the same name.
          type: array
          maxItems: 100
          uniqueItems: true
          items: { $ref: "#/components/schemas/SecretName" }
        # http or https, with a host and without credentials or a fragment.
        callback_url:
          description: Receives a signed POST when a task is retried or finishes.
          type: string
          maxLength: 2048
        in_process:
          description: Run the concurrency slots as threads of one runner process.
          type: boolean
          default: false
        lifecycle_hooks:
          $ref: "#/components/schemas/LifecycleHooks"
        client_contract:
          description: The signature `lazycloud app export` types clients from.
          type: object
          additionalProperties: true
          x-go-type: json.RawMessage
        http:
          $ref: "#/components/schemas/HttpSpec"
    SourceRef:
      type: object
      required: [sha256]
      properties:
        sha256: { $ref: "#/components/schemas/Sha256" }
    ImageSpec:
      type: object
      required: [python_version]
      properties:
        python_version:
          description: The Python minor version the runtime is mounted for.
          type: string
          enum: ["3.10", "3.11", "3.12", "3.13", "3.14"]
        # A ready image the workspace resolved, built for python_version.
        # Without one the function runs the platform's image for it.
        image_id: { $ref: "#/components/schemas/ImageId" }
        reference:
          description: The image by digest the release runs, set by the server.
          type: string
          readOnly: true
    Resources:
      description: >-
        Reservations the container always keeps. CPU above the reservation is
        shared up to `cpu_limit_millis`, by default the reservation plus 16
        cores. Memory above the reservation is allowed up to
        `memory_limit_mib`, by default four times the reservation, at least
        1 GiB and at most 8 GiB above it; the container is killed beyond it.
      type: object
      required: [cpu_millis, memory_mib]
      properties:
        cpu_millis:
          type: integer
          minimum: 125
          maximum: 192000
        cpu_limit_millis:
          type: integer
          minimum: 125
          maximum: 192000
        memory_mib:
          type: integer
          minimum: 128
          maximum: 1572864
        memory_limit_mib:
          type: integer
          minimum: 128
          maximum: 1572864
    RetryPolicy:
      type: object
      required: [max_attempts]
      properties:
        max_attempts:
          description: Attempts including the first.
          type: integer
          minimum: 1
          maximum: 100
        delay_seconds:
          type: number
          minimum: 0
          maximum: 3600
          default: 0
        backoff:
          type: string
          enum: [fixed, exponential]
          default: fixed
        max_delay_seconds:
          type: number
          minimum: 0
          maximum: 86400
    Autoscaler:
      type: object
      properties:
        min_containers:
          type: integer
          minimum: 0
          maximum: 1000
          default: 0
        max_containers:
          type: integer
          minimum: 1
          maximum: 1000
          default: 1
        tasks_per_container:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1
    Release:
      type: object
      required: [id, function, created_at, spec]
      properties:
        id: { type: string, format: uuid }
        function: { $ref: "#/components/schemas/WorkloadName" }
        version:
          description: The deployed version; absent for a release only working-tree calls use.
          type: integer
        created_at: { type: string, format: date-time }
        spec: { $ref: "#/components/schemas/FunctionSpec" }
        url:
          description: Where an HTTP workload answers, following the active release.
          type: string
...
    TaskStatus:
      type: string
      enum: [queued, running, succeeded, failed, cancelled]
    Task:
      type: object
      required: [id, app, function, release_id, status, attempts, max_attempts, created_at, root_task_id]
      properties:
        id: { type: string, format: uuid }
        app: { $ref: "#/components/schemas/AppName" }
        function: { $ref: "#/components/schemas/WorkloadName" }
        release_id: { type: string, format: uuid }
        version:
          description: The deployed version of the release; absent for a working-tree release.
          type: integer
        status: { $ref: "#/components/schemas/TaskStatus" }
        attempts:
          description: Attempts started so far.
          type: integer
        max_attempts: { type: integer }
        parent_task_id:
          description: The running task that spawned this one.
          type: string
          format: uuid
        root_task_id:
          description: The root of the call graph; the task itself when nothing spawned it.
          type: string
          format: uuid
        scheduled_for:
          description: The cron occurrence that admitted the task.
          type: string
          format: date-time
        container_id:
          description: The container of the latest attempt.
          type: string
          format: uuid
        next_attempt_at:
          description: When a queued task that already ran becomes due again.
          type: string
          format: date-time
        pending: { $ref: "#/components/schemas/TaskPendingProgress" }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        failure: { $ref: "#/components/schemas/TaskFailure" }
    TaskPendingProgress:
      description: Why a queued task has not started, derived from current state on every read.
      type: object
      required: [reason, message, since, pending_since, observed_at]
      properties:
        reason: { $ref: "#/components/schemas/TaskPendingReason" }
        message: { type: string }
        since:
          description: When the current reason began.
          type: string
          format: date-time
        pending_since:
          description: When the task last became due or was submitted.
          type: string
          format: date-time
        observed_at: { type: string, format: date-time }
    TaskPendingReason:
      type: string
      enum:
        - queued
        - dependencies
        - retry
        - capacity_busy
        - capacity_unavailable
        - capacity_limit
        - provisioning_compute
        - starting_container
      x-enum-varnames:
        - Queued
        - Dependencies
        - Retry
        - CapacityBusy
        - CapacityUnavailable
        - CapacityLimit
        - ProvisioningCompute
        - StartingContainer
    FailureKind:
      type: string
      enum: [user_error, load_error, timeout, lost, start_failed, system, dependency_failed]
    TaskFailure:
      type: object
      required: [kind, message]
      properties:
        kind: { $ref: "#/components/schemas/FailureKind" }
        type:
          description: The exception type the handler raised.
          type: string
        message: { type: string }
        traceback: { type: string }
        exception:
          description: The cloudpickled exception, when the runner could serialize it.
          type: string
          format: byte
...
    LifecycleHooks:
      description: >-
        Callables the runner invokes with a context object, each a
        `module:qualname` reference into the source. on_start runs once per
        runner process after the handler loads, and its failure is a load
        error. The others run in the container around each attempt, in
        order; their failures are logged and do not change the outcome.
      type: object
      properties:
        on_start: { $ref: "#/components/schemas/HookReferences" }
        on_running: { $ref: "#/components/schemas/HookReferences" }
        on_success: { $ref: "#/components/schemas/HookReferences" }
        on_error: { $ref: "#/components/schemas/HookReferences" }
        on_retry: { $ref: "#/components/schemas/HookReferences" }
        on_failure: { $ref: "#/components/schemas/HookReferences" }
        on_finish: { $ref: "#/components/schemas/HookReferences" }
    HookReferences:
      type: array
      maxItems: 16
      items:
        type: string
        pattern: "^[A-Za-z_][A-Za-z0-9_.]*:[A-Za-z_][A-Za-z0-9_.]*$"
    SecretName:
      description: An environment variable name; the LAZYCLOUD_ prefix is reserved.
      type: string
      pattern: "^[A-Za-z_][A-Za-z0-9_]{0,239}$"
...
    ImageId:
      type: string
      pattern: "^img_[0-9a-f]{24}$"
...
    HttpSpec:
      description: >-
        Makes the workload answer HTTP. Its slots are `workers` processes,
        each handling `concurrency` requests at once.
      type: object
      required: [kind]
      properties:
        kind: { $ref: "#/components/schemas/HttpKind" }
        route:
          description: The path an endpoint answers on.
          type: string
          pattern: "^/[^\\s?#]*$"
          maxLength: 512
          default: /
        methods:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/HttpMethod" }
          default: [GET, POST]
        domain:
          $ref: "#/components/schemas/Hostname"
        authorized:
          description: Require a bearer token for the workspace.
          type: boolean
          default: true
        workers:
          type: integer
          minimum: 1
          maximum: 64
          default: 1
    HttpKind:
      type: string
      enum: [endpoint, asgi, realtime]
    HttpMethod:
      type: string
      enum: [DELETE, GET, HEAD, OPTIONS, PATCH, POST, PUT, TRACE]
    Hostname:
      type: string
      maxLength: 253
      pattern: "^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?(?:\\.[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?)+$"
    HttpWorkload:
      type: object
      required: [name, app, kind, state, release, url, version_url, release_url]
      properties:
        name: { $ref: "#/components/schemas/WorkloadName" }
        app: { $ref: "#/components/schemas/AppName" }
        kind: { $ref: "#/components/schemas/HttpKind" }
        state:
          type: string
          enum: [active, stopped]
        release:
          $ref: "#/components/schemas/Release"
        url:
          description: Follows the active release across deploys.
          type: string
        version_url:
          description: Pinned to the release's version.
          type: string
        release_url:
          description: Addresses the release by id.
          type: string
        domain_url:
          description: The custom hostname, once its registration is ready.
          type: string
    InvocationBody:
      type: object
      additionalProperties: true
      # Kept as raw JSON so arguments pass through unchanged, including
      # integers beyond float64 precision.
      x-go-type: json.RawMessage
    Invocation:
      type: object
      required: [task]
      properties:
        task: { $ref: "#/components/schemas/Task" }
        result:
          description: The JSON value the function returned, once it succeeded.
          x-go-type: json.RawMessage
    DomainRequest:
      type: object
      required: [hostname]
      properties:
        hostname:
          type: string
          maxLength: 253
    DomainPhase:
      type: string
      enum: [awaiting_verification, validating, ready, action_required]
    DomainErrorCode:
      type: string
      enum: [verification_timed_out, certificate_failed, hostname_rejected, upstream_unavailable]
    DnsRecord:
      type: object
      required: [type, name, value]
      properties:
        type: { type: string }
        name: { type: string }
        value: { type: string }
    Domain:
      type: object
      required: [id, hostname, phase, cname_target, required_records, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        hostname: { $ref: "#/components/schemas/Hostname" }
        phase: { $ref: "#/components/schemas/DomainPhase" }
        cname_target:
          description: The hostname the domain's CNAME points at.
          type: string
        required_records:
          description: Records the provider still waits for beyond the CNAME.
          type: array
          items: { $ref: "#/components/schemas/DnsRecord" }
        error_code: { $ref: "#/components/schemas/DomainErrorCode" }
        error_message: { type: string }
        verified_at: { type: string, format: date-time }
        last_checked_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DomainList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Domain" }
        next:
          description: Pass as `after` for the next page; absent on the last.
          type: string
...